Express: README and basic layout

+++ laravel/notejam/app/controllers/PadController.php
@@ -2,47 +2,44 @@
 
 class PadController extends BaseController {
 
-    public function create()
-    {
-        return View::make('pad/create');
-    }
-
-    public function store()
-    {
-        $validation = $this->validator();
-        if ($validation->fails())
+	public function create()
+	{
+        if (Request::isMethod('post'))
         {
-            return Redirect::route('pads.create')->withErrors($validation);
+            $validation = $this->validator();
+            if ($validation->fails())
+            {
+                return Redirect::route('create_pad')->withErrors($validation);
+            }
+            $pad = new Pad(array('name' => Input::get('name')));
+            Auth::user()->pads()->save($pad);
+            return Redirect::route('view_pad', array('id' => $pad->id))
+                ->with('success', 'Pad is created.');
         }
-        $pad = new Pad(array('name' => Input::get('name')));
-        Auth::user()->pads()->save($pad);
-        return Redirect::route('pads.show', array('id' => $pad->id))
-            ->with('success', 'Pad is created.');
-    }
+		return View::make('pad/create');
+	}
 
-    public function edit($id)
-    {
+	public function edit($id)
+	{
         $pad = $this->getPadOrFail($id);
-        return View::make('pad/edit', array('pad' => $pad));
-    }
-
-    public function update($id)
-    {
-        $pad = $this->getPadOrFail($id);
-        $validation = $this->validator();
-        if ($validation->fails())
+        if (Request::isMethod('post'))
         {
-            return Redirect::route('pads.edit', array('id' => $pad->id))
-                ->withErrors($validation);
+            $validation = $this->validator();
+            if ($validation->fails())
+            {
+                return Redirect::route('edit_pad', array('id' => $pad->id))
+                    ->withErrors($validation);
+            }
+            $pad->name = Input::get('name');
+            $pad->save();
+            return Redirect::route('view_pad', array('id' => $pad->id))
+                ->with('success', 'Pad is updated.');
         }
-        $pad->name = Input::get('name');
-        $pad->save();
-        return Redirect::route('pads.show', array('id' => $pad->id))
-            ->with('success', 'Pad is updated.');
-    }
+		return View::make('pad/edit', array('pad' => $pad));
+	}
 
-    public function show($id)
-    {
+	public function view($id)
+	{
         $pad = $this->getPadOrFail($id);
         $orderParams = $this->processOrderParam();
         $notes = $pad->notes()->orderBy(
@@ -52,21 +49,19 @@ class PadController extends BaseController {
             'pad/view',
             array('pad' => $pad, 'notes' => $notes)
         );
-    }
+	}
 
-    public function delete($id)
-    {
+	public function delete($id)
+	{
         $pad = $this->getPadOrFail($id);
-        return View::make('pad/delete', array('pad' => $pad));
-    }
-
-    public function destroy($id)
-    {
-        $pad = $this->getPadOrFail($id);
-        $pad->delete();
-        return Redirect::route('all_notes')
-            ->with('success', 'Pad is deleted.');
-    }
+        if (Request::isMethod('post'))
+        {
+            $pad->delete();
+            return Redirect::route('all_notes')
+                ->with('success', 'Pad is deleted.');
+        }
+		return View::make('pad/delete', array('pad' => $pad));
+	}
 
     private function getPadOrFail($id)
     {
@@ -83,5 +78,6 @@ class PadController extends BaseController {
             )
         );
     }
+
 }
